overviewGarminRunning

A function to generate stylish plots that rapidly give an overview of your running.

overviewGarminRunning(
  data,
  date_from = NA,
  plot,
  target_time = median(data$Avg.Pace)
)

Arguments

date_from
a date processed by as.Date e.g as.Date("2020-03-18"). This is a cut-off data for your analysis. Everything after this will be included in the analysis.

plot
can be one of "runs_per_loc_per_dist", "distance_per_location", "mean_pace", "pca_plot", and "total_runs"

target_time
a pace in the form "7:30" (character variable) with "minutes:seconds" that is one previously achieved that you wish to mark on your plots, OR it could be a target pace you have set yourself

output
from the processGarminRunning function

Value

plot

Examples

#What about my pace - calculate my mean avg.pace amongst all my runs per location

overviewGarminRunning(my_runs,
                     date_from = as.Date("2020-03-18"),
                     plot = "mean_pace",
                     target_time = "7:18")

#> Loading required package: tidyverse

#> ── Attaching packages ─────────────────────────────────────── tidyverse 1.3.0 ──

#> ✔ ggplot2 3.3.2     ✔ purrr   0.3.4
#> ✔ tibble  3.0.3     ✔ dplyr   1.0.2
#> ✔ tidyr   1.1.1     ✔ stringr 1.4.0
#> ✔ readr   1.3.1     ✔ forcats 0.5.0

#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ dplyr::filter() masks stats::filter()
#> ✖ dplyr::lag()    masks stats::lag()

#> Loading required package: lubridate

#> 
#> Attaching package: ‘lubridate’

#> The following objects are masked from ‘package:base’:
#> 
#>     date, intersect, setdiff, union

#> Loading required package: RColorBrewer

#> Loading required package: ggridges

#> Loading required package: ggpubr

#> Loading required package: ggfortify

#> Loading required package: caret

#> Loading required package: lattice

#> 
#> Attaching package: ‘caret’

#> The following object is masked from ‘package:purrr’:
#> 
#>     lift

#> Error in overviewGarminRunning(my_runs, date_from = as.Date("2020-03-18"),     plot = "mean_pace", target_time = "7:18"): object 'my_runs' not found

#Note the 5-10 mile runs the pace is slightly faster in manchester than in Bromsgrove probably because of the hills!

# we can also utilise numeric data to do PCA analysis - what are the major sources of variation in our running data?
overviewGarminRunning(my_runs,
                     date_from = as.Date("2020-03-18"),
                     plot = "pca_plot",
                     target_time = "7:18")

#> Error in overviewGarminRunning(my_runs, date_from = as.Date("2020-03-18"),     plot = "pca_plot", target_time = "7:18"): object 'my_runs' not found